# docs/api/swagger.yaml
openapi: 3.0.3
info:
  title: Salon Management API
  version: 1.0.0
  description: Combined API documentation for Salon Management System (Auth, Branches, Categories, Salons, Staff)
servers:
  - url: http://localhost:3000/api/v1
    description: Local Development Server
...
paths:
...
  /staff:
    post:
      tags: [Staff]
      summary: Create staff
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffRequest"
      responses:
        "201":
          description: Staff created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessStaffResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

    get:
      tags: [Staff]
      summary: List staff by salon
      security:
        - bearerAuth: []
      parameters:
        - name: salon_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: inactive
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Staff list fetched successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessStaffListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /staff/{id}:
    get:
      tags: [Staff]
      summary: Get staff by id
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/StaffIdParam"
      responses:
        "200":
          description: Staff fetched successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessStaffResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      tags: [Staff]
      summary: Update staff
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/StaffIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStaffRequest"
      responses:
        "200":
          description: Staff updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessStaffResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [Staff]
      summary: Delete staff
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/StaffIdParam"
      responses:
        "200":
          description: Staff deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessDeleteResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ===================== STAFF SCHEDULES =====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    StaffIdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Validation error / Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing/invalid access token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Validation error"
        code:
          type: string
          example: "VALIDATION_ERROR"
...
    SuccessDeleteResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            id:
              type: string
              format: uuid
...
    Staff:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        salon_id:
          type: string
          format: uuid
        branch_id:
          type: string
          format: uuid
          nullable: true
        designation:
          type: string
          nullable: true
        specialization:
          type: array
          items:
            type: string
        experience_years:
          type: integer
          nullable: true
        commission_type:
          type: string
          nullable: true
        commission_value:
          type: number
          nullable: true
        is_active:
          type: boolean
        joined_date:
          type: string
          format: date
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateStaffRequest:
      type: object
      required: [user_id, salon_id]
      properties:
        user_id:
          type: string
          format: uuid
        salon_id:
          type: string
          format: uuid
        branch_id:
          type: string
          format: uuid
        designation:
          type: string
        specialization:
          type: array
          items:
            type: string
        experience_years:
          type: integer
        commission_type:
          type: string
        commission_value:
          type: number
        is_active:
          type: boolean
        joined_date:
          type: string
          format: date
    UpdateStaffRequest:
      type: object
      description: Partial update
      properties:
        branch_id:
          type: string
          format: uuid
        designation:
          type: string
        specialization:
          type: array
          items:
            type: string
        experience_years:
          type: integer
        commission_type:
          type: string
        commission_value:
          type: number
        is_active:
          type: boolean
        joined_date:
          type: string
          format: date
    SuccessStaffResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "OK"
        data:
          $ref: "#/components/schemas/Staff"
    SuccessStaffListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "OK"
        data:
          type: array
          items:
            $ref: "#/components/schemas/Staff"
...